5_file_storage/requirements.txt
aiofiles==23.1.0
alembic==1.10.4
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
dnspython==2.3.0
docker==6.1.2
dockerpty==0.4.1
fastapi==0.95.1
fastapi-cache==0.1.0
fastapi-redis-cache==0.2.5
fastapi-users==11.0.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
jsonschema==3.2.0
loguru==0.7.0
orjson==3.8.12
packaging==23.1
path==16.6.0
pendulum==2.1.2
pydantic==1.10.7
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
redis==4.3.5
requests==2.30.0
SQLAlchemy==2.0.13
SQLAlchemy-Utils==0.41.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
Werkzeug==2.3.4
